feat(BoM): introducing the Spark Bill of Materials
And rename `SparkAndroidPublishingPlugin` into `SparkPublishingPlugin`.
