+++ build-logic/src/main/kotlin/com/adevinta/spark/ProjectExtensions.kt
@@ -41,6 +41,7 @@ internal val Project.isAndroidApplication: Boolean get() = pluginManager.hasPlug
 internal val Project.isAndroidLibrary: Boolean get() = pluginManager.hasPlugin("com.android.library")
 internal val Project.isAndroidTest: Boolean get() = pluginManager.hasPlugin("com.android.test")
 internal val Project.isAndroid: Boolean get() = pluginManager.hasPlugin("com.android.base")
+internal val Project.isJavaPlatform: Boolean get() = pluginManager.hasPlugin("org.gradle.java-platform")
 
 internal fun Project.configureAndroidExtension(
     configure: CommonExtension<*, *, *, *, *>.() -> Unit,
