+++ settings.gradle.kts
@@ -44,7 +44,8 @@ enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 rootProject.name = "spark-android"
 
 include(":spark")
-include(":sample")
+include(":spark-bom")
 include(":spark-icons")
 include(":spark-lint")
 include(":spark-screenshot-testing")
+include(":sample")
